Replace PromptInput with LandingPageInput and extract shared types
- Create LandingPageInput component for landing page only (simplified from 450 to 130 lines)
- Extract MODELS, ModelId, StreamPhase, GenerationErrorType to types/generation.ts
- Delete PromptInput.tsx (generation logic now lives in ChatSidebar + useGenerationApi)
- Fix ChatInput to use Enter to submit (Shift+Enter for newlines)
- Update all imports to use new type locations
